+++ SimFolk/src/sk/upjs/ics/mmizak/simfolk/core/database/access/services/implementations/SongService.java
@@ -20,7 +20,7 @@ public class SongService implements ISongService {
     }
 
     @Override
-    public Song getById(Integer id) {
+    public Song getById(Long id) {
         return songDao.getById(id);
     }
